sentiment_analysis/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 25.3.1
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
annotated-types=0.7.0=pypi_0
anthropic=0.69.0=pypi_0
anyio=4.10.0=py312h06a4308_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=25.1.0=py312hee96239_0
asttokens=3.0.0=pyhd8ed1ab_1
async-lru=2.0.4=py312h06a4308_0
attrs=24.3.0=py312h06a4308_0
babel=2.16.0=py312h06a4308_0
beautifulsoup4=4.13.5=py312h06a4308_0
bleach=6.2.0=py312h06a4308_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.9.9=h06a4308_0
certifi=2025.8.3=pypi_0
cffi=1.17.1=py312h1fdaa30_1
charset-normalizer=3.4.3=pypi_0
comm=0.2.3=pyhe01879c_0
debugpy=1.8.16=py312hbdd6827_0
decorator=5.2.1=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd3eb1b0_0
distro=1.9.0=pypi_0
docstring-parser=0.17.0=pypi_0
exceptiongroup=1.3.0=pyhd8ed1ab_0
executing=2.2.1=pyhd8ed1ab_0
expat=2.7.1=h6a678d5_0
filelock=3.13.1=pypi_0
fsspec=2024.6.1=pypi_0
h11=0.16.0=py312h06a4308_0
hf-xet=1.1.10=pypi_0
httpcore=1.0.9=py312h06a4308_0
httpx=0.28.1=py312h06a4308_0
huggingface-hub=0.35.0=pypi_0
idna=3.10=pypi_0
importlib-metadata=8.7.0=pyhe01879c_1
ipykernel=6.30.1=pyh82676e8_0
ipython=9.5.0=pyhfa0c392_0
ipython_pygments_lexers=1.1.1=pyhd8ed1ab_0
ipywidgets=8.1.5=py312h06a4308_0
jedi=0.19.2=pyhd8ed1ab_1
jinja2=3.1.4=pypi_0
jiter=0.11.0=pypi_0
json5=0.9.25=py312h06a4308_0
jsonschema=4.25.0=py312h06a4308_0
jsonschema-specifications=2023.7.1=py312h06a4308_0
jupyter=1.1.1=py312h06a4308_0
jupyter-lsp=2.2.5=py312h06a4308_0
jupyter_client=8.6.3=pyhd8ed1ab_1
jupyter_console=6.6.3=py312h06a4308_1
jupyter_core=5.8.1=pyh31011fe_0
jupyter_events=0.12.0=py312h06a4308_0
jupyter_server=2.16.0=py312h06a4308_0
jupyter_server_terminals=0.5.3=py312h06a4308_0
jupyterlab=4.4.7=py312h06a4308_0
jupyterlab_pygments=0.3.0=py312h06a4308_0
jupyterlab_server=2.27.3=py312h06a4308_0
jupyterlab_widgets=3.0.15=py312h06a4308_0
ld_impl_linux-64=2.40=h12ee557_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libsodium=1.0.20=heac8642_0
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
libxcb=1.17.0=h9b100fa_0
libzlib=1.3.1=hb25bd0a_0
markupsafe=2.1.5=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_1
mistune=3.1.2=py312h06a4308_0
mpmath=1.3.0=pypi_0
nbclient=0.10.2=py312h06a4308_0
nbconvert=7.16.6=py312h06a4308_0
nbconvert-core=7.16.6=py312h06a4308_0
nbconvert-pandoc=7.16.6=py312h06a4308_0
nbformat=5.10.4=py312h06a4308_0
ncurses=6.5=h7934f7d_0
nest-asyncio=1.6.0=pyhd8ed1ab_1
networkx=3.3=pypi_0
notebook=7.4.5=py312h06a4308_0
notebook-shim=0.2.4=py312h06a4308_0
numpy=2.1.2=pypi_0
nvidia-cublas-cu12=12.8.4.1=pypi_0
nvidia-cuda-cupti-cu12=12.8.90=pypi_0
nvidia-cuda-nvrtc-cu12=12.8.93=pypi_0
nvidia-cuda-runtime-cu12=12.8.90=pypi_0
nvidia-cudnn-cu12=9.10.2.21=pypi_0
nvidia-cufft-cu12=11.3.3.83=pypi_0
nvidia-cufile-cu12=1.13.1.3=pypi_0
nvidia-curand-cu12=10.3.9.90=pypi_0
nvidia-cusolver-cu12=11.7.3.90=pypi_0
nvidia-cusparse-cu12=12.5.8.93=pypi_0
nvidia-cusparselt-cu12=0.7.1=pypi_0
nvidia-nccl-cu12=2.27.3=pypi_0
nvidia-nvjitlink-cu12=12.8.93=pypi_0
nvidia-nvtx-cu12=12.8.90=pypi_0
openssl=3.0.17=h5eee18b_0
overrides=7.4.0=py312h06a4308_0
packaging=25.0=pyh29332c3_1
pandas=2.3.2=pypi_0
pandoc=3.8=h06a4308_0
pandocfilters=1.5.1=py312h06a4308_0
parso=0.8.5=pyhcf101f3_0
pexpect=4.9.0=pyhd8ed1ab_1
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=11.0.0=pypi_0
pip=25.2=pyhc872135_0
platformdirs=4.4.0=pyhcf101f3_0
prometheus_client=0.21.1=py312h06a4308_0
prompt-toolkit=3.0.43=py312h06a4308_0
prompt_toolkit=3.0.43=hd3eb1b0_0
psutil=7.0.0=py312hee96239_0
pthread-stubs=0.3=h0ce48e5_1
ptyprocess=0.7.0=pyhd8ed1ab_1
pure_eval=0.2.3=pyhd8ed1ab_1
pyarrow=21.0.0=pypi_0
pycparser=2.23=py312h06a4308_0
pydantic=2.11.9=pypi_0
pydantic-core=2.33.2=pypi_0
pygments=2.19.2=pyhd8ed1ab_0
python=3.12.11=h22baa00_0
python-dateutil=2.9.0.post0=pyhe01879c_2
python-fastjsonschema=2.20.0=py312h06a4308_0
python-json-logger=3.2.1=py312h06a4308_0
pytz=2025.2=pypi_0
pyyaml=6.0.2=pypi_0
pyzmq=27.1.0=py312hcf8288c_0
qtconsole=5.7.0=py312h06a4308_0
qtpy=2.4.3=py312h06a4308_0
readline=8.3=hc2a1206_0
referencing=0.30.2=py312h06a4308_0
regex=2025.9.18=pypi_0
requests=2.32.5=pypi_0
rfc3339-validator=0.1.4=py312h06a4308_0
rfc3986-validator=0.1.1=py312h06a4308_0
rpds-py=0.22.3=py312h4aa5aa6_0
safetensors=0.6.2=pypi_0
send2trash=1.8.2=py312h06a4308_1
setuptools=78.1.1=py312h06a4308_0
six=1.17.0=pyhe01879c_1
sniffio=1.3.0=py312h06a4308_0
soupsieve=2.5=py312h06a4308_0
sqlite=3.50.2=hb25bd0a_1
stack_data=0.6.3=pyhd8ed1ab_1
sympy=1.13.3=pypi_0
terminado=0.18.1=py312h06a4308_0
tinycss2=1.4.0=py312h06a4308_0
tk=8.6.15=h54e0aa7_0
tokenizers=0.22.1=pypi_0
torch=2.8.0+cu128=pypi_0
torchvision=0.23.0+cu128=pypi_0
tornado=6.5.1=py312h5eee18b_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pyhd8ed1ab_1
transformers=4.56.2=pypi_0
triton=3.4.0=pypi_0
typing-extensions=4.12.2=pypi_0
typing-inspection=0.4.2=pypi_0
typing_extensions=4.15.0=pyhcf101f3_0
tzdata=2025.2=pypi_0
urllib3=2.5.0=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_1
webencodings=0.5.1=py312h06a4308_2
websocket-client=1.8.0=py312h06a4308_0
wheel=0.45.1=py312h06a4308_0
widgetsnbextension=4.0.13=py312h06a4308_0
xorg-libx11=1.8.12=h9b100fa_1
xorg-libxau=1.0.12=h9b100fa_0
xorg-libxdmcp=1.1.5=h9b100fa_0
xorg-xorgproto=2024.1=h5eee18b_1
xz=5.6.4=h5eee18b_1
zeromq=4.3.5=hb0a5e54_1
zipp=3.23.0=pyhd8ed1ab_0
zlib=1.3.1=hb25bd0a_0
